Rename layout_builder to executive_summary.
