+++ xcp_d/interfaces/__init__.py
@@ -6,8 +6,8 @@ from xcp_d.interfaces import (
     bids,
     c3,
     connectivity,
+    executive_summary,
     filtering,
-    layout_builder,
     nilearn,
     plotting,
     prepostcleaning,
@@ -24,7 +24,7 @@ __all__ = [
     "c3",
     "connectivity",
     "filtering",
-    "layout_builder",
+    "executive_summary",
     "nilearn",
     "prepostcleaning",
     "plotting",
