+++ xcp_d/interfaces/executive_summary.py
